# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Nimpass REST API
  version: 1.0.0
  license: { name: MIT, url: 'https://spdx.org/licenses/MIT.html' }
  description: |
    All money values are integer Luna. Login and payout proofs must be produced
    by Nimiq Pay's sign(message), using the exact server-provided challenge message.
    Payout completion requires signatures by both the new payout wallet and the
    authenticated owner wallet over that same challenge message.
    Unsafe methods require an exact configured Origin (or Referer fallback).
    Shared middleware errors use the Error schema: ORIGIN_FORBIDDEN (403),
    AUTH_REQUIRED (401), CSRF_INVALID (403), CONTENT_TYPE (415),
    VALIDATION_ERROR (400), RATE_LIMITED (429) and INTERNAL_ERROR (500).
    Session lookup infrastructure failures return INTERNAL_ERROR (500), preserving
    the cookie; AUTH_REQUIRED (401) means a missing, invalid, expired or revoked session.
    Authenticated operations require a session; every unsafe authenticated
    operation requires CSRF. JSON bodies reject unknown fields and trailing data.
    Text length bounds are UTF-8 byte limits after documented normalization.
    Authenticated unsafe methods additionally require X-CSRF-Token, obtained
    from login or GET /auth/session. Cookies are HttpOnly, SameSite=Lax; HTTPS
    deployments use Secure __Host-nimpass_session, local HTTP development uses
    nimpass_session. Requests must send credentials: include.
...
    unchanged.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /passes/{passID}/redemption-challenges:
    parameters: [ { $ref: '#/components/parameters/PassID' } ]
    post:
      summary: Create the current customer redemption challenge
      operationId: createRedemptionChallenge
      description: Creates a five-minute, Pass/provider/customer-bound challenge and returns the exact canonical message to sign. Repeating the call while one is live returns that same challenge. Creating a challenge never consumes a session.
      parameters: [ { $ref: '#/components/parameters/Csrf' } ]
      responses:
        default: { $ref: '#/components/responses/Error' }
        '200': { description: Current or newly created challenge, content: { application/json: { schema: { $ref: '#/components/schemas/RedemptionChallenge' } } } }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '410': { $ref: '#/components/responses/Error' }
...
  /passes/{passID}/redemptions:
    parameters: [ { $ref: '#/components/parameters/PassID' } ]
    get:
      summary: List own Pass redemption history
      operationId: listPassRedemptions
      responses:
        default: { $ref: '#/components/responses/Error' }
        '200': { description: Customer-owned redemption history, content: { application/json: { schema: { $ref: '#/components/schemas/RedemptionHistory' } } } }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /redemption-challenges/{challengeID}:
    parameters: [ { $ref: '#/components/parameters/ChallengeID' } ]
    get:
      summary: Get own redemption challenge state
      operationId: getRedemptionChallenge
      responses:
        default: { $ref: '#/components/responses/Error' }
        '200': { description: Customer-owned challenge, content: { application/json: { schema: { $ref: '#/components/schemas/RedemptionChallenge' } } } }
        '404': { $ref: '#/components/responses/Error' }
  /redemption-challenges/{challengeID}/authorization:
    parameters: [ { $ref: '#/components/parameters/ChallengeID' } ]
    post:
      summary: Authorize a redemption challenge with the customer wallet
      operationId: authorizeRedemption
      description: 'Verifies one fresh Nimiq Pay AUTHORIZE_REDEMPTION signature over the server-authored canonical message and consumes exactly one session in the same transaction. The Pass owner''s signature is the sole authority: there is no second-party confirmation step. Expiry, Pass binding, Pass status and stale context are all re-checked under row locks before the session is spent, and a replayed authorization consumes nothing.'
      parameters: [ { $ref: '#/components/parameters/Csrf' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RedemptionAuthorization' } } }
      responses:
        default: { $ref: '#/components/responses/Error' }
        '200': { description: Consumed challenge with the authoritative remaining count, content: { application/json: { schema: { $ref: '#/components/schemas/RedemptionChallenge' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '410': { $ref: '#/components/responses/Error' }
  /providers/{providerID}/redemptions:
    parameters: [ { $ref: '#/components/parameters/ProviderID' } ]
    get:
      summary: List redemption history for an owned Provider
      operationId: listProviderRedemptions
      responses:
        default: { $ref: '#/components/responses/Error' }
        '200': { description: Provider-owned operational history, content: { application/json: { schema: { $ref: '#/components/schemas/RedemptionHistory' } } } }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-nimpass_session
      description: HTTPS production cookie; local HTTP development uses nimpass_session.
  parameters:
    Csrf: { in: header, name: X-CSRF-Token, required: true, schema: { type: string, pattern: '^[0-9a-f]{64}$' } }
    ProviderID: { in: path, name: providerID, required: true, schema: { type: string, format: uuid } }
...
    PassID: { in: path, name: passID, required: true, schema: { type: string, format: uuid } }
...
    ChallengeID: { in: path, name: challengeID, required: true, schema: { type: string, format: uuid } }
...
  responses:
    Error:
      description: Normalized error; codes include AUTH_REQUIRED, INVALID_SIGNATURE, CHALLENGE_EXPIRED, CHALLENGE_CONSUMED, FORBIDDEN, VALIDATION_ERROR, NOT_FOUND, CONFLICT, PAYMENT_CONFLICT, PASS_UNAVAILABLE, PASS_ALREADY_OWNED, PURCHASE_IN_SETTLEMENT, PASS_PURCHASE_CUTOFF, SELF_PURCHASE_NOT_ALLOWED, INTENT_EXPIRED, PASS_NOT_FOUND, PASS_NOT_OWNED, PASS_EXPIRED, PASS_COMPLETED, REDEMPTION_CHALLENGE_EXPIRED, REDEMPTION_ALREADY_CONSUMED, REDEMPTION_NOT_AUTHORIZED, INVALID_REDEMPTION_SIGNATURE, STALE_REDEMPTION_CHALLENGE, INVALID_REDEMPTION_TOKEN, RATE_LIMITED, CSRF_INVALID, ORIGIN_FORBIDDEN.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
...
    PassSessionStatus:
      type: string
      enum: [UNSCHEDULED, SCHEDULED, COMPLETED, CANCELLED]
      description: >-
        UNSCHEDULED has no date, SCHEDULED has one, and both are open. COMPLETED is the only status that spends
        a session; it and CANCELLED are terminal. A completed session keeps whatever date it had.
    PassSession:
      type: object
      required: [id, passId, sequenceNumber, status, scheduledAt, completedAt, completedBy, redemptionId, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        passId: { type: string, format: uuid, description: The purchased Pass this session belongs to. }
        sequenceNumber: { type: integer, format: int32, minimum: 1, description: '1-based position, fixed at purchase and never renumbered.' }
        status: { $ref: '#/components/schemas/PassSessionStatus' }
        scheduledAt: { type: string, format: date-time, nullable: true }
        completedAt: { type: string, format: date-time, nullable: true }
        completedBy: { type: string, enum: [OWNER, PROVIDER], nullable: true, description: Who recorded the delivery. OWNER means a wallet-signed redemption spent it. }
        redemptionId: { type: string, format: uuid, nullable: true, description: 'The signed redemption that spent this session, when there was one.' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    RedemptionAuthorization:
      type: object
      additionalProperties: false
      required: [publicKey, signature]
      properties:
        publicKey: { type: string, pattern: '^[0-9a-fA-F]{64}$' }
        signature: { type: string, pattern: '^[0-9a-fA-F]{128}$' }
        signingScheme: { $ref: '#/components/schemas/SigningScheme' }
    RedemptionChallenge:
      type: object
      required: [challengeId, passId, providerId, purpose, status, message, createdAt, expiresAt, authorizedAt, consumedAt, pass, redemption, session]
      properties:
        challengeId: { type: string, format: uuid }
        passId: { type: string, format: uuid }
        providerId: { type: string, format: uuid }
        purpose: { type: string, enum: [AUTHORIZE_REDEMPTION] }
        status: { type: string, enum: [CREATED, AUTHORIZED, CONSUMED, EXPIRED, CANCELLED] }
        message: { type: string, description: Exact canonical message to pass to Nimiq Pay sign(). }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        authorizedAt: { type: string, format: date-time, nullable: true }
        consumedAt: { type: string, format: date-time, nullable: true }
        pass: { type: object, required: [status, originalSessions, usedSessions, remainingSessions, expiresAt], properties: { status: { type: string, enum: [ACTIVE, COMPLETED, EXPIRED, CANCELLED] }, originalSessions: { type: integer }, usedSessions: { type: integer }, remainingSessions: { type: integer }, expiresAt: { type: string, format: date-time, nullable: true } } }
        redemption: { type: object, nullable: true, properties: { id: { type: string, format: uuid }, sessionOrdinal: { type: integer }, redeemedAt: { type: string, format: date-time } } }
        session:
          type: object
          nullable: true
          allOf: [ { $ref: '#/components/schemas/PassSession' } ]
          description: The Pass session this authorization spent. Present only on a successful authorization, so the screen can name which session moved rather than only that the count changed.
    RedemptionHistory:
      type: object
      required: [items]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/RedemptionHistoryItem' } }
    RedemptionHistoryItem:
      type: object
      required: [redemptionId, challengeId, passId, providerId, serviceId, sourcePassId, sessionOrdinal, ownerWallet, redeemedAt, status]
      properties:
        redemptionId: { type: string, format: uuid }
        challengeId: { type: string, format: uuid }
        passId: { type: string, format: uuid }
        providerId: { type: string, format: uuid }
        serviceId: { type: string, format: uuid }
        sourcePassId: { type: string, format: uuid }
        sessionOrdinal: { type: integer, minimum: 1 }
        ownerWallet: { type: string }
        redeemedAt: { type: string, format: date-time }
        status: { type: string, enum: [CONSUMED] }
    SigningScheme:
      type: string
      enum: [raw, hub]
      description: >-
        Which documented preprocessing the wallet applied before signing.
        Omit it to verify under the deployment's configured NIMIQ_SIGNING_SCHEME.
        Send "hub" for a signature produced by @nimiq/hub-api signMessage(),
        whose envelope is specified as sha256("\x16Nimiq Signed Message:\n" +
        length + message). The Nimiq Pay Mini App provider does not document its
        preprocessing, so that transport sends nothing. Exactly one scheme is
        applied per proof; the server never retries under another.
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }